working on moving Listing details and index into routes, update Listings store
